fixed save bug with temporary particles. added duration argument to /h set

+++ ParticleHats/src/com/mediusecho/particlehats/util/ResourceUtil.java
@@ -142,7 +142,7 @@ public class ResourceUtil {
 	
 	public static InputStream getMostCompatibleParticlesMenu ()
 	{
-		int menuVersion = particleMenuCompatibilityMap.getOrDefault(Math.min(ParticleHats.serverVersion, 14), 8);
+		int menuVersion = particleMenuCompatibilityMap.getOrDefault(Math.min((int) ParticleHats.serverVersion, 14), 8);
 		return core.getResource("menus/particles_" + menuVersion + ".yml");
 	}
 }
